[Feat] added jwt and validations in controllers

+++ server/src/controllers/user.controller.ts
@@ -2,18 +2,38 @@ import { Request, Response } from 'express';
 import bcrypt from 'bcrypt';
 import { validationResult } from 'express-validator';
 import fs from 'fs-extra';
+import jwt from 'jsonwebtoken';
+import validator from 'validator';
 
 import User, { IUser } from '../models/user';
 
 // get all users
 export async function getUsers(req: Request, res: Response): Promise<Response> {
-  const users = await User.find().select('-password');
-  
-  return res.json(users);
+  if(!validator.isMongoId(req.user.id)) {
+    return res.status(400).json({ msg: 'is not valid id' });
+  }
+
+  const user = await User.findById(req.user.id);
+  if(!user) {
+    return res.status(403).json({ msg: 'invalid permissions' })
+  }
+
+  try {
+    const users = await User.find().select('-password');
+    return res.json(users);
+  } catch(err) {
+    return res.status(500).json({ msg: 'there was an error' });
+  }
+
 }
 
 // get one user by id
 export async function getUser(req: Request, res: Response): Promise<Response> {
+  const errors = validationResult(req);
+  if(!errors.isEmpty()) {
+    return res.status(400).json(errors);
+  }
+  
   const user = await User.findById(req.params.id);
   
   return res.json(user);
@@ -50,11 +70,23 @@ export async function createUser(req: Request, res: Response): Promise<Response>
 
   try {
     await user.save();
+
+    const payload = {
+      user: {
+        id: user._id
+      }
+    }
+
+    const token = jwt.sign(payload, process.env.SECRET as string, {
+      expiresIn: 86400
+    });
+
+    return res.json(token);
+
   } catch(err) {
     return res.status(500).json({ msg: 'there was a error' });
   }
  
-  return res.json(user);
 }
 
 export async function updateUser(req: Request, res: Response): Promise<Response> {
@@ -66,6 +98,9 @@ export async function updateUser(req: Request, res: Response): Promise<Response>
   }
 
   let user = await User.findById(req.params.id) as IUser;
+  if(user._id !== req.user.id) {
+    return res.status(403).json({ msg: 'not authorized' });
+  }
 
   const { password } = req.body;
   const salt = await bcrypt.genSalt();
@@ -83,7 +118,20 @@ export async function updateUser(req: Request, res: Response): Promise<Response>
 // delete user
 export async function deleteUser(req: Request, res: Response): Promise<Response> {
 
-  const user = await User.findByIdAndDelete(req.params.id) as IUser;
+  const errors = validationResult(req)
+  if(!errors.isEmpty()) {
+    return res.status(400).json(errors);
+  }
+
+  let user = await User.findById(req.user.id) as IUser;
+  if(req.user.id !== user._id ) {
+    return res.status(403).json({ msg: 'not authorized' });
+  }
+
+  user = await User.findByIdAndDelete(req.params.id) as IUser;
+  if(!user) {
+    return res.status(404).json({ msg: 'user not exist' });
+  }
   await fs.unlink(user.avatar);
 
   return res.json(user);
